split out metrics
