+++ allennlp/training/metrics/__init__.py
@@ -1,0 +1,5 @@
+from allennlp.training.metrics.metric import Metric
+from allennlp.training.metrics.boolean_accuracy import BooleanAccuracy
+from allennlp.training.metrics.categorical_accuracy import CategoricalAccuracy
+from allennlp.training.metrics.f1_measure import F1Measure
+from allennlp.training.metrics.span_based_f1_measure import SpanBasedF1Measure
