+++ allennlp/training/metrics/categorical_accuracy.py
@@ -1,0 +1,82 @@
+from typing import Optional
+
+from overrides import overrides
+import torch
+from torch.autograd import Variable
+
+from allennlp.common.checks import ConfigurationError
+from allennlp.training.metrics.metric import Metric
+
+
+@Metric.register("categorical_accuracy")
+class CategoricalAccuracy(Metric):
+    """
+    Categorical Top-K accuracy. Assumes integer labels, with
+    each item to be classified having a single correct class.
+    """
+    def __init__(self, top_k: int = 1) -> None:
+        self._top_k = top_k
+        self.correct_count = 0.
+        self.total_count = 0.
+
+    def __call__(self,
+                 predictions: torch.Tensor,
+                 gold_labels: torch.Tensor,
+                 mask: Optional[torch.Tensor] = None):
+        """
+        Parameters
+        ----------
+        predictions : ``torch.Tensor``, required.
+            A tensor of predictions of shape (batch_size, ..., num_classes).
+        gold_labels : ``torch.Tensor``, required.
+            A tensor of integer class label of shape (batch_size, ...). It must be the same
+            shape as the ``predictions`` tensor without the ``num_classes`` dimension.
+        mask: ``torch.Tensor``, optional (default = None).
+            A masking tensor the same size as ``gold_labels``.
+        """
+        # If you actually passed in Variables here instead of Tensors, this will be a huge memory
+        # leak, because it will prevent garbage collection for the computation graph.  We'll ensure
+        # that we're using tensors here first.
+        if isinstance(predictions, Variable):
+            predictions = predictions.data
+        if isinstance(gold_labels, Variable):
+            gold_labels = gold_labels.data
+        if isinstance(mask, Variable):
+            mask = mask.data
+
+        # Some sanity checks.
+        num_classes = predictions.size(-1)
+        if gold_labels.dim() != predictions.dim() - 1:
+            raise ConfigurationError("gold_labels must have dimension == predictions.size() - 1 but "
+                                     "found tensor of shape: {}".format(predictions.size()))
+        if (gold_labels >= num_classes).any():
+            raise ConfigurationError("A gold label passed to Categorical Accuracy contains an id >= {}, "
+                                     "the number of classes.".format(num_classes))
+
+        # Top K indexes of the predictions (or fewer, if there aren't K of them)
+        top_k = predictions.topk(min(self._top_k, predictions.shape[-1]), -1)[1]
+
+        # This is of shape (batch_size, ..., top_k).
+        correct = top_k.eq(gold_labels.long().unsqueeze(-1)).float()
+        count = torch.ones(gold_labels.size())
+        if mask is not None:
+            correct *= mask.unsqueeze(-1)
+            count *= mask
+        self.correct_count += correct.sum()
+        self.total_count += count.sum()
+
+    def get_metric(self, reset: bool = False):
+        """
+        Returns
+        -------
+        The accumulated accuracy.
+        """
+        accuracy = float(self.correct_count) / float(self.total_count)
+        if reset:
+            self.reset()
+        return accuracy
+
+    @overrides
+    def reset(self):
+        self.correct_count = 0.0
+        self.total_count = 0.0
