+++ allennlp/training/metrics/f1_measure.py
@@ -1,0 +1,103 @@
+from typing import Optional
+
+import torch
+from torch.autograd import Variable
+
+from allennlp.training.metrics.metric import Metric
+from allennlp.common.checks import ConfigurationError
+
+
+@Metric.register("f1")
+class F1Measure(Metric):
+    """
+    Computes Precision, Recall and F1 with respect to a given ``positive_label``.
+    For example, for a BIO tagging scheme, you would pass the classification index of
+    the tag you are interested in, resulting in the Precision, Recall and F1 score being
+    calculated for this tag only.
+    """
+    def __init__(self, positive_label: int) -> None:
+        self._positive_label = positive_label
+        self._true_positives = 0.0
+        self._true_negatives = 0.0
+        self._false_positives = 0.0
+        self._false_negatives = 0.0
+
+    def __call__(self,
+                 predictions: torch.Tensor,
+                 gold_labels: torch.Tensor,
+                 mask: Optional[torch.Tensor] = None):
+        """
+        Parameters
+        ----------
+        predictions : ``torch.Tensor``, required.
+            A tensor of predictions of shape (batch_size, ..., num_classes).
+        gold_labels : ``torch.Tensor``, required.
+            A tensor of integer class label of shape (batch_size, ...). It must be the same
+            shape as the ``predictions`` tensor without the ``num_classes`` dimension.
+        mask: ``torch.Tensor``, optional (default = None).
+            A masking tensor the same size as ``gold_labels``.
+        """
+        # If you actually passed in Variables here instead of Tensors, this will be a huge memory
+        # leak, because it will prevent garbage collection for the computation graph.  We'll ensure
+        # that we're using tensors here first.
+        if isinstance(predictions, Variable):
+            predictions = predictions.data
+        if isinstance(gold_labels, Variable):
+            gold_labels = gold_labels.data
+        if isinstance(mask, Variable):
+            mask = mask.data
+
+        num_classes = predictions.size(-1)
+        if (gold_labels >= num_classes).any():
+            raise ConfigurationError("A gold label passed to F1Measure contains an id >= {}, "
+                                     "the number of classes.".format(num_classes))
+        if mask is None:
+            mask = torch.ones(gold_labels.size())
+        mask = mask.float()
+        gold_labels = gold_labels.float()
+        positive_label_mask = gold_labels.eq(self._positive_label).float()
+        negative_label_mask = 1.0 - positive_label_mask
+
+        argmax_predictions = predictions.topk(1, -1)[1].float().squeeze(-1)
+
+        # True Negatives: correct non-positive predictions.
+        correct_null_predictions = (argmax_predictions !=
+                                    self._positive_label).float() * negative_label_mask
+        self._true_negatives += (correct_null_predictions.float() * mask).sum()
+
+        # True Positives: correct positively labeled predictions.
+        correct_non_null_predictions = (argmax_predictions ==
+                                        self._positive_label).float() * positive_label_mask
+        self._true_positives += (correct_non_null_predictions * mask).sum()
+
+        # False Negatives: incorrect negatively labeled predictions.
+        incorrect_null_predictions = (argmax_predictions !=
+                                      self._positive_label).float() * positive_label_mask
+        self._false_negatives += (incorrect_null_predictions * mask).sum()
+
+        # False Positives: incorrect positively labeled predictions
+        incorrect_non_null_predictions = (argmax_predictions ==
+                                          self._positive_label).float() * negative_label_mask
+        self._false_positives += (incorrect_non_null_predictions * mask).sum()
+
+    def get_metric(self, reset: bool = False):
+        """
+        Returns
+        -------
+        A tuple of the following metrics based on the accumulated count statistics:
+        precision : float
+        recall : float
+        f1-measure : float
+        """
+        precision = float(self._true_positives) / float(self._true_positives + self._false_positives + 1e-13)
+        recall = float(self._true_positives) / float(self._true_positives + self._false_negatives + 1e-13)
+        f1_measure = 2. * ((precision * recall) / (precision + recall + 1e-13))
+        if reset:
+            self.reset()
+        return precision, recall, f1_measure
+
+    def reset(self):
+        self._true_positives = 0.0
+        self._true_negatives = 0.0
+        self._false_positives = 0.0
+        self._false_negatives = 0.0
